Code inconsistence corrected.

+++ src/main/java/com/dub/spring/site/actors/ActorIdForm.java
@@ -9,16 +9,14 @@ public class ActorIdForm {
 
 	@Min(value  = 1, message = "{validate.min.actorId}")
 	@NotNull(message = "{validate.required.actorId}")
-	Integer id;
+	long id;
 
-	public Integer getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(long id) {
 		this.id = id;
 	}
-	
-	
 	
 }
